requirements.txt
# ==============================================================================
# REPORTSMITH - CORE DEPENDENCIES
# ==============================================================================

# Configuration & Settings
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# Database Connectivity
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
cx_Oracle>=8.3.0
pyodbc>=4.0.0

# Async Database Support
asyncpg>=0.28.0
aiopg>=1.4.0

# AI & Machine Learning
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Financial Data APIs
yfinance>=0.2.0
alpha-vantage>=2.3.0
quandl>=3.7.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# ==============================================================================
# OPTIONAL: Data Ingestion Library (Install separately if needed)
# ==============================================================================
# For local development:
# pip install -e ../ReportSmith-DataIngestion
# 
# For production:
# reportsmith-data-ingestion>=1.0.0
